+++ jest.config.ts
@@ -21,6 +21,10 @@ const config: Config.InitialOptions = {
       "<rootDir>/packages/implementations/react/src/atomic/Box/src",
     "@rarui-react/button":
       "<rootDir>/packages/implementations/react/src/atomic/Button/src",
+    "@rarui-react/chip":
+      "<rootDir>/packages/implementations/react/src/atomic/Chip/src",
+    "@rarui-react/checkbox":
+      "<rootDir>/packages/implementations/react/src/atomic/Checkbox/src",
     "@rarui-react/link":
       "<rootDir>/packages/implementations/react/src/atomic/Link/src",
     "@rarui-react/text":