+++ packages/implementations/react/src/atomic/Select/src/index.ts
@@ -1,0 +1,5 @@
+import { Select } from "./Select";
+
+export { Select } from "./Select";
+export type { SelectProps } from "./select.types";
+export default Select;