+++ packages/implementations/react/src/index.ts
@@ -14,6 +14,7 @@ export * from "@rarui-react/label";
 export * from "@rarui-react/link";
 export * from "@rarui-react/progress";
 export * from "@rarui-react/radio-button";
+export * from "@rarui-react/select";
 export * from "@rarui-react/skeleton";
 export * from "@rarui-react/status";
 export * from "@rarui-react/text";