+++ packages/styles/src/packages/atomic/chip/rarui-chip.css.ts
@@ -21,6 +21,7 @@ const chip = recipe({
     position: "relative",
     overflow: "hidden",
     borderStyle: "solid",
+    transition: "all 150ms ease",
     gap: varsThemeBase.spacing["3xs"],
     fontWeight: varsThemeBase.fontWeight.medium,
     fontFamily: varsThemeBase.fontFamily.inter,
@@ -29,6 +30,10 @@ const chip = recipe({
     borderColor: varsThemeBase.colors.border.subdued,
     backgroundColor: varsThemeBase.colors.surface.primary,
     color: varsThemeBase.colors.content.primary,
+    ":focus-within": {
+      outline: "none",
+      boxShadow: varsThemeBase.utils["focus-ring"],
+    },
     ":disabled": {
       background: varsThemeBase.colors.surface.disabled,
       borderColor: varsThemeBase.colors.surface.disabled,
@@ -83,6 +88,19 @@ const chip = recipe({
   },
 });
 
+export const close = style({
+  transition: "all 150ms ease",
+  color: varsThemeBase.colors.content.secondary,
+  borderRadius: varsThemeBase.shape.border.radius.pill,
+  ":focus-within": {
+    outline: "none",
+    boxShadow: varsThemeBase.utils["focus-ring"],
+  },
+  ":hover": {
+    color: varsThemeBase.colors.content.primary,
+  },
+});
+
 export const overlay = style({
   position: "absolute",
   inset: 0,
@@ -101,6 +119,7 @@ globalStyle(`${chip()}:active ${overlay}`, {
 
 export const styles = {
   chip,
+  close,
   overlay,
 };