+++ packages/styles/src/packages/atomic/select/index.ts
@@ -1,0 +1,6 @@
+import { selectStyles } from "./rarui-select.css";
+
+export const select = {
+  classnames: selectStyles,
+};
+export type { SelectVariants } from "./rarui-select.types";