+++ packages/styles/src/packages/atomic/select/rarui-select.types.ts
@@ -1,0 +1,6 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { selectStyles } from "./rarui-select.css";
+
+export type SelectVariants = NonNullable<
+  RecipeVariants<typeof selectStyles.select>
+>;