+++ packages/implementations/react/src/composite/Dropdown/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Dropdown are informative, specific, and action -oriented text labels that provide contextual support.
 
+## 2024-06-12 `1.1.0`
+
+#### 🎉 New features
+
+- Added `enabledClick` property to the Component. ([#57](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/57) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-06-02 `1.0.1`
 
 #### 💡 Others